remove multi_buffer. corrected tests
