+++ autotune/core/utils.py
@@ -262,47 +262,6 @@ def matmul_block(lhsT_block, rhs_block, result_block):
             result_block[tile_id_M, tile_id_N, idx_res.p, idx_res.x] += result_tile[idx_res.p, idx_res.x]
 
 
-def save_result_block(result, result_block, m_ofs: int, n_ofs: int):
-    """
-    Store result_block into result
-    Args:
-    result: M, N
-    result_block: TILES_IN_BLOCK_M, TILES_IN_BLOCK_N, nl.par_dim(TILE_M), TILE_N
-    """
-    TILES_IN_BLOCK_M, TILES_IN_BLOCK_N, TILE_M, TILE_N = result_block.shape
-
-    idx_res = nl.mgrid[0:TILE_M, 0:TILE_N]
-    for tile_id_M in nl.affine_range(TILES_IN_BLOCK_M):
-        m_start = m_ofs + tile_id_M * TILE_M
-        for tile_id_N in nl.affine_range(TILES_IN_BLOCK_N):
-            n_start = n_ofs + tile_id_N * TILE_N
-            nl.store(result[m_start + idx_res.p, n_start + idx_res.x], value=result_block[tile_id_M, tile_id_N])
-
-
-def save_result_acc(result, result_tiles, BLOCK_M, BLOCK_N):
-    NUM_BLOCK_K, NUM_BLOCK_M, NUM_BLOCK_N, num_m_tiles, num_n_tiles, TILE_M, TILE_N = result_tiles.shape
-
-    idx_res = nl.mgrid[0:TILE_M, 0:TILE_N]
-    for block_id_M in nl.affine_range(NUM_BLOCK_M):
-        m_ofs = block_id_M * BLOCK_M
-        for block_id_N in nl.affine_range(NUM_BLOCK_N):
-            n_ofs = block_id_N * BLOCK_N
-            for tile_id_M in nl.affine_range(num_m_tiles):
-                for tile_id_N in nl.affine_range(num_n_tiles):
-                    result_acc = nl.zeros(
-                        (num_m_tiles, num_n_tiles, nl.par_dim(TILE_M), TILE_N), dtype=result_tiles.dtype, buffer=nl.sbuf
-                    )
-                    for block_id_K in nl.affine_range(NUM_BLOCK_K):
-                        result_acc[tile_id_M, tile_id_N] += result_tiles[
-                            block_id_K, block_id_M, block_id_N, tile_id_M, tile_id_N
-                        ]
-
-                    nl.store(
-                        result[m_ofs + tile_id_M * TILE_M + idx_res.p, n_ofs + tile_id_N * TILE_N + idx_res.x],
-                        value=result_acc[tile_id_M, tile_id_N],
-                    )
-
-
 def transpose_tiles_in_block(block):
     """
     Transpose the (pmax, pmax) tiles in a block in place
